fix: test error GitHub CI
